requirements.txt
Django==1.6.3
PyYAML==3.11
South==0.8.4
dj-database-url==0.3.0
django-cors-headers==0.12
django-filter==0.7
django-rest-swagger==0.1.14
djangorestframework==2.3.13
gevent==1.0
gnureadline==6.3.3
greenlet==0.4.2
gunicorn==18.0
hiredis==0.1.3
ipython==2.0.0
psycogreen==1.0
psycopg2==2.5.2
redis==2.9.1
requests==2.2.1
wsgiref==0.1.2
